backend/requirements.txt
# Core API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# LLM / AI
groq>=0.9.0
sentence-transformers>=2.7.0
faiss-cpu>=1.9.0
rank-bm25>=0.2.2

# OCR
easyocr>=1.7.0

# RAG / NLP
langchain-text-splitters>=0.2.0

# HTTP / Scraping
httpx>=0.27.0
aiohttp>=3.10.0
beautifulsoup4>=4.12.0
lxml>=5.3.0

# Data
pydantic>=2.7.0
python-dotenv>=1.0.0
numpy>=2.0.0
pillow>=10.3.0

# Utilities
tenacity>=8.3.0
loguru>=0.7.0
tqdm>=4.66.0
